+++ test/mocha/30-blocks-api.js
@@ -64,5 +64,34 @@ describe('Blocks API', () => {
         done();
       });
     });
+    describe('getLatestBlockHeight', () => {
+      it('gets the blockHeigh without useCache option', async () => {
+        let error;
+        let result;
+        try {
+          result = await ledgerNode.blocks.getLatestBlockHeight();
+        } catch(e) {
+          error = e;
+        }
+        assertNoError(error);
+        should.exist(result);
+        result.should.equal(0);
+
+      });
+      it('should get a blockHeight with useCache option', async () => {
+        let error;
+        let result;
+        try {
+          result = await ledgerNode.blocks.getLatestBlockHeight(
+            {useCache: true});
+        } catch(e) {
+          error = e;
+        }
+        assertNoError(error);
+        should.exist(result);
+        result.should.equal(0);
+      });
+
+    }); // end getLatestBlockHeight
   });
 });